Authentication (#3)
* create auth module

* validate AuthCredentialsDto

* handle duplicate user

* install bcrypt

* encrypt user password

* allow api to sign in a user

* install passport related libraries

* do a dumb configuration in auth module

* return a jwt from user authentication

* add types for passport-jwt

* create jwt strategy

* add a custom getUser decorator

* make tasks accessible by creating user only

* configure unit tests

* create new isNonemptyArray helper

* getAllTask route now return tasks related to the requesting user only

* getTaskById now get a specific task providen the tasks exist and the user is the owner

* allow deletion of task to owner of the task user

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
+import isNonemptyArray from 'src/helpers/isNonemptyArray/isNonemptyArray';
 
 @Injectable()
 export class TasksService {
@@ -13,30 +15,36 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    const tasks = await this.taskRepository.getTasks(filterDto, user);
+
+    if (!isNonemptyArray(tasks)) throw new NotFoundException("User doesn't register tasks");
+
+    return tasks;
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
     if (!found) throw new NotFoundException(`Task with ID: '${id}' not found`);
     return found;
   }
 
-  async deleteTask(id: number): Promise<{ message: string }> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<{ message: string }> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (!result.affected) throw new NotFoundException(`Task with ID: '${id}' not found`);
     return { message: `Task with ID: ${id} deleted` };
   }
-  async updateTaskStatus(args: { id: number; status: TaskStatus }): Promise<Task> {
-    const { id, status } = args;
-    const task = await this.getTaskById(id);
+
+  async updateTaskStatus(args: { id: number; status: TaskStatus; user: User }): Promise<Task> {
+    const { id, status, user } = args;
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
 
     return task;
   }
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 }
